src/segmentation/requirement.txt
absl-py==2.3.1
ackermann-msgs==2.0.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.8.0
ament-cmake-test==1.3.11
ament-copyright==0.12.12
ament-cppcheck==0.12.12
ament-cpplint==0.12.12
ament-flake8==0.12.12
ament-index-python==1.4.0
ament-lint==0.12.12
ament-lint-cmake==0.12.12
ament-package==0.14.0
ament-pep257==0.12.12
ament-uncrustify==0.12.12
ament-xmllint==0.12.12
angles==1.15.0
antlr4-python3-runtime==4.9.3
black==25.9.0
builtin-interfaces==1.2.1
click==8.3.0
cloudpickle==3.1.1
composition-interfaces==1.2.1
contourpy==1.3.2
control-msgs==4.8.0
controller-manager-msgs==2.50.0
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.8.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.3
examples-rclpy-minimal-action-client==0.15.3
examples-rclpy-minimal-action-server==0.15.3
examples-rclpy-minimal-client==0.15.3
examples-rclpy-minimal-publisher==0.15.3
examples-rclpy-minimal-service==0.15.3
examples-rclpy-minimal-subscriber==0.15.3
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
fvcore==0.1.5.post20221221
gazebo-msgs==3.9.0
generate-parameter-library-py==0.4.0
geometry-msgs==4.8.0
grpcio==1.75.1
hydra-core==1.3.2
image-geometry==3.2.1
interactive-markers==2.3.2
iopath==0.1.9
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
kiwisolver==1.4.9
laser-geometry==2.4.0
launch==1.0.8
launch-param-builder==0.1.1
launch-ros==0.19.9
launch-testing==1.0.8
launch-testing-ros==0.19.9
launch-xml==1.0.8
launch-yaml==1.0.8
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.10.7
message-filters==4.3.7
mocap4r2-control-msgs==0.0.7
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
mpmath==1.3.0
mypy_extensions==1.1.0
nav-msgs==4.8.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
omegaconf==2.3.0
opencv-python==4.12.0.88
osrf-pycommon==2.1.6
packaging==25.0
pathspec==0.12.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.3.0
platformdirs==4.5.0
portalocker==3.2.0
protobuf==6.33.0
pycocotools==2.0.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytokens==0.2.0
PyYAML==6.0.3
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.16
rcutils==5.1.6
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.12
ros2bag==0.15.14
ros2cli==0.18.12
ros2component==0.18.12
ros2doctor==0.18.12
ros2interface==0.18.12
ros2launch==0.19.9
ros2lifecycle==0.18.12
ros2multicast==0.18.12
ros2node==0.18.12
ros2param==0.18.12
ros2pkg==0.18.12
ros2run==0.18.12
ros2service==0.18.12
ros2topic==0.18.12
rosbag2-interfaces==0.15.14
rosbag2-py==0.15.14
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.6
rosidl-cli==3.1.6
rosidl-cmake==3.1.6
rosidl-generator-c==3.1.6
rosidl-generator-cpp==3.1.6
rosidl-generator-py==0.14.4
rosidl-parser==3.1.6
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.6
rosidl-typesupport-introspection-cpp==3.1.6
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-console==2.0.3
rqt-gauges==0.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.45.0
rqt-moveit==1.0.1
rqt-msg==1.2.0
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.1
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.5
rqt-topic==1.5.0
sensor-msgs==4.8.0
sensor-msgs-py==4.8.0
shape-msgs==4.8.0
six==1.17.0
srdfdom==2.0.7
sros2==0.10.6
statistics-msgs==1.2.1
std-msgs==4.8.0
std-srvs==4.8.0
stereo-msgs==4.8.0
sympy==1.14.0
tabulate==0.9.0
teleop-twist-keyboard==2.4.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf2-geometry-msgs==0.25.12
tf2-kdl==0.25.12
tf2-msgs==0.25.12
tf2-py==0.25.12
tf2-ros-py==0.25.12
tf2-tools==0.25.12
tomli==2.3.0
topic-monitor==0.20.5
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
trajectory-msgs==4.8.0
triton==3.3.1
turtlesim==1.4.2
typeguard==4.4.4
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
visualization-msgs==4.8.0
Werkzeug==3.1.3
xacro==2.0.13
yacs==0.1.8
